@@ -14,24 +14,25 @@ gemspec
 # gem 'debugger'
 
 # Set the version of Rails for the dummy app
-gem 'rails', '~> 5.2.6'
+gem 'rails', '~> 7.0.2'
 
 # Lock in Rake to a version compatible with rspec-rails 3.0
-gem 'rake', '< 12.0.0'
+gem 'rake'
 
 gem 'sass', require: 'sass'
 
 group :test, :development do
-  gem 'rspec-rails', '~> 3.6.0'
+  gem 'rspec-rails'
   gem 'pry'
 end
 
 group :test do
-  gem 'factory_girl_rails', '~> 4.4.1'
-  gem 'capybara-webkit', '~> 1.11.1'
+  gem 'webrick'
+  gem 'factory_girl_rails'
+  gem 'capybara-webkit'
   gem 'capybara-screenshot'
-  gem 'guard-rspec', '~> 4.7.3'
-  gem 'selenium-webdriver', '~> 2.42.0'
+  gem 'guard-rspec'
+  gem 'selenium-webdriver'
   gem 'shoulda-matchers', require: false
   gem 'yarjuf'
   gem 'database_cleaner'
