@@ -125,7 +125,7 @@ module Fae
 
     def load_and_filter_models
       # load of all models since Rails caches activerecord queries.
-      Rails.application.eager_load!
+      #Rails.application.eager_load!
       ActiveRecord::Base.descendants.map.reject { |m| m.name['Fae::'] || !m.instance_methods.include?(:fae_display_field) || Fae.dashboard_exclusions.include?(m.name) || !authorize_model(m) }
     end
 
