@@ -1,3 +1,5 @@
-Judge.configure do
-  expose Fae::User, :email
+Rails.application.config.to_prepare do
+  Judge.configure do
+    expose Fae::User, :email
+  end
 end
