@@ -1,5 +1,11 @@
-Fae::StaticPage.class_eval do
-  def instance_is_decorated
-    "Fae::StaticPage instance is decorated"
+class StaticPageDecorator
+
+  def initialize
+    Fae::StaticPage.class_eval do
+      def instance_is_decorated
+        "Fae::StaticPage instance is decorated"
+      end
+    end
   end
+
 end
