@@ -1,7 +1,9 @@
 FactoryGirl.define do
 
   factory :fae_image, class: 'Fae::Image' do
-    asset { Rack::Test::UploadedFile.new(File.join('spec', 'support', 'assets', 'test.jpg')) }
+    # https://github.com/docker/for-mac/issues/5570
+    file = File.open(File.join('spec', 'support', 'assets', 'test.jpg'))
+    asset { Rack::Test::UploadedFile.new(file, 'image/jpeg', true, original_filename: 'test.jpg') }
   end
 
 end
